Implement recursive pydantic_to_marshmallow transformation, refactoring
